MAINT: Split stratigraphic framework for easier reuse
