+++ frontend/src/components/project/stratigraphicFramework/FrameworkData.ts
@@ -1,0 +1,5 @@
+import { createContext } from "react";
+
+import type { FrameworkData } from "./types";
+
+export const FrameworkDataContext = createContext<FrameworkData | null>(null);
