+++ frontend/src/components/project/stratigraphicFramework/utils.ts
@@ -1,0 +1,5 @@
+import type { ItemType } from "./types";
+
+export function findIndexByName(items: ItemType[], name: string) {
+  return items.findIndex((item) => item.name === name);
+}
